Two conflicts, both where the two lines of work had solved the same
problem:

- gobj.c: both sides tightened the render-callback guard. Kept
  pc_code_ptr_ok() (the module's own executable PT_LOAD, found through
  dl_iterate_phdr) over the __executable_start/etext form, because the
  port is a shared library on Android where those symbols describe the
  wrong module. The reason the other version existed is the sharper half
  of the rationale, so it moved into pc_ptr.h: bounding to the executable
  segment is what rejects a pointer into .rodata, which is exactly what
  the results screen supplies.

- lb/types.h: the bg-flash work reintroduced
  PC_SCRIPT_BE = scalar_storage_order("big-endian") for the colour-overlay
  script word, and its definition landed first in the merged file -- so it
  would have applied to all 84 command structs that Phase 1 rewrote for
  the host word, misreading every fighter script, and Clang rejects the
  attribute outright. PC_SCRIPT_BE is empty again, defined before its
  first user, and union ColorOverlay_x8_t is declared for the host word
  like the rest. That is not a behaviour change: lb_013B.c already calls
  pc_script_prepare() on this very script, so its words arrive byteswapped
  and the attribute would have converted them twice.

Desktop golden suite within baseline; the Android arm64 library links,
which the attribute alone would have prevented.

+++ extern/dolphin/src/dolphin/ax/AXAlloc.c
@@ -71,7 +71,7 @@ AXVPB* __AXPopFreeStack(void)
 {
     AXVPB* p;
 
-    p = (void*) (u32) &__AXStackHead[0]->next;
+    p = (void*) (uintptr_t) &__AXStackHead[0]->next;
     if (p) {
         __AXStackHead[0] = p->next;
     }
@@ -92,7 +92,7 @@ AXVPB* __AXPopCallbackStack(void)
 {
     AXVPB* p;
 
-    p = (void*) (u32) &__AXCallbackStack[0];
+    p = (void*) (uintptr_t) &__AXCallbackStack[0];
     if (p) {
         __AXCallbackStack = p->next1;
     }
